\name{FishFC}
\docType{data}
\alias{FishFC}
\encoding{latin1}
\title{FACTORS AND FRESHWATER FISH SPECIES}
\usage{data(FishFC)}
\description{Relative contribution of factors to explain the distribution of some freshwater fish species of the orders Amiiformes, Atheriniformes and Perciformes. The file was obtained using ModestR.} 
\format{An data frame with 22 columns: taxonomic data (order, family, genus and species) and the contribution of 17 factors.}
\source{The records of the species and some environmental variables such as altitude, aspect, slope, topographic heterogeneity (TH24), terrestrial primary production (TPP), vegetation index (VI) and population density (Pop) were obtained from Guisande et al. (2010) and Pelayo-Villamil et al. (2015). The variables mean annual temperature (BIO1), temperature diurnal range (BIO2), isothermality (BIO3), temperature seasonality (BIO4), mean temperature of wettest quarter (BIO8), annual precipitation (BIO12), precipitation of driest month (BIO14), precipitation seasonality (BIO15), precipitation of warmest quarter (BIO18) and precipitation of coldest quarter (BIO19) were downloaded from the web \url{https://www.worldclim.org/}.}
\references{
Guisande, C., Manjarrés-Hernández, A., Pelayo-Villamil, P., Granado-Lorencio, C., Riveiro, I., Acuña, A., Prieto-Piraquive, E., Janeiro, E., Matías, J.M., Patti, C., Patti, B., Mazzola, S., Jiménez, S., Duque, V. & Salmerón, F. (2010) IPez: An expert system for the taxonomic identification of fishes based on machine learning techniques. \emph{Fisheries Research}, 102, 240-247.

Hijmans, R.J., Cameron, S.E., Parra, J.L., Jones, P.G. and Jarvis, A. (2005) Very high resolution interpolated climate surfaces for global land areas. \emph{International Journal of Climatology}, 25, 1965-1978.

Pelayo-Villamil, P., Guisande, C., Vari, R.P., Manjarrés-Hernández, A., García-Roselló, E., González-Dacosta, J. et al. (2015) Global diversity patterns of freshwater fishes - potential victims of their own success. \emph{Diversity and Distributions}, 21, 345-356.
}
\keyword{FishFC}
