\name{Instability}
\docType{data}
\alias{Instability}
\encoding{latin1}
\title{INSTABILITY INDEX AND FREQUENCIES FOR PRESENCES AND THE EXTENT}
\usage{data(Instability)}
\description{Instability index of some freshwater fish species of the orders Amiiformes and Perciformes. The file was obtained using ModestR (Pelayo-Villamil et al., 2012; Garcia-Roselló et al., 2013).} 
\format{An data frame with 15 columns: taxonomic data (order, family, genus and species), variable, number of interval, value of the interval, frequencies of the cells of presence and the extent, frequencies standardized to 0-1 for both the cells of presence and the extent, instability, instability standardized to 0-1 and the contribution of the factor in percentage.}
\source{As mentioned in \link{FishFC}.}
\keyword{Instability}
