\name{VI}
\docType{data}
\alias{VI}
\title{CONTRIBUTION OF VEGETATION INDEX}
\description{Relative contribution of vegetation index to explain species distribution of freshwater fish species, in cells of 1 degree around the world.}
\usage{data(VI)}
\format{A matrix with 180 rows (latitudes) and 360 columns (longitudes).}
\keyword{VI}
